chore: refactor thunks for moveLbalesForRbf
